pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.12.37",
    "llama-index-embeddings-huggingface>=0.5.4",
    "llama-index-llms-lmstudio>=0.3.0",
    "llama-index-llms-ollama>=0.5.6",
    "llama-index-readers-file>=0.4.7",
    "python-dotenv>=1.1.0",
]
